4UCON 3.5mm terminal blocks
Based on TerminalBlock_Pheonix_PT-3.5mm_2pol.kicad_mod

The 4UCON terminal blocks are extremely similar to the Pheonix ones,
but they stick out a little less in front and a little more in back.

Datasheet is here:
https://cdn-shop.adafruit.com/datasheets/19963.pdf

(19963 is the "drawing number".  Not sure if that is the part number,
but don't know what else to use as the part number.)

Generated with this script:
https://github.com/pointhi/kicad-footprint-generator/blob/master/scripts/Connectors_Terminal_Blocks/terminal_block.py